refactored debug id into cli-config
