+++ workspaces/cli-config/package.json
@@ -18,7 +18,9 @@
     "find-up": "^4.1.0",
     "get-port": "^5.1.0",
     "js-yaml": "^3.13.1",
-    "path-to-regexp": "^3.0.0"
+    "nice-try": "^2.1.0",
+    "path-to-regexp": "^3.0.0",
+    "human-readable-ids": "^1.0.4"
   },
   "devDependencies": {
     "@types/find-up": "^4.0.0",
