+++ workspaces/cli-server/src/server.ts
@@ -26,7 +26,7 @@ import { analyticsEventEmitter, track } from './analytics';
 import cors from 'cors';
 import { IgnoreFileHelper } from '@useoptic/cli-config/build/helpers/ignore-file-interface';
 import { Session, SessionsManager } from './sessions';
-import { getOrCreateAnonId } from '@useoptic/cli/lib/shared/authentication-server';
+import { getOrCreateAnonId } from '@useoptic/cli-config/build/opticrc/optic-rc';
 
 const pJson = require('../package.json');
