+++ workspaces/local-cli/src/commands/debug/identity.ts
@@ -1,5 +1,5 @@
 import { Command } from '@oclif/command';
-import { getOrCreateAnonId } from '../../shared/authentication-server';
+import { getOrCreateAnonId } from '@useoptic/cli-config/build/opticrc/optic-rc';
 export default class DebugIdentity extends Command {
   static description =
     'returns the anonymous id that identifies diff engine logs in Sentry';
